Add ReduceDatatypeClasses helper method
Can optionally be called from the OnBeforeProcessParseTree callback, to remove classes that inherit from e.g. %String, and are merely present to host parameters such as MAXLEN or XMLPATTERN. This helper moves the parameters to the properties themselves, and deletes these classes. On some XSDs, this may significantly reduce the number of generated classes.