Add VSCodeCompat post-processor
The Visual Studio Code objectscript plugin has a bug that causes the final newline of an export to be removed. The code owners refuse to fix this bug, for backwards compatibility reasons; see

https://github.com/intersystems-community/vscode-objectscript/issues/850

This postprocessor can be used for repositories where the exports need to be compatible with the objectscript plugin.