Add support for manual-only tests
Sometimes the test machinery is used for reproducing a specific issue; in this case it is inconvenient if this test is included in the All Tests or Package runs. A parameter can now be used to exclude it; it will then only run if the class is specified explicitly to the runner.